OpenWebNet: change wait time between energy totalizer request (#2104)
* OpenWebNet: bug fix and burglar alarm battery, network and maintenance status and fix warnings

Fix wrong automatic creation of auxiliary device and update status of
dry contact, introduced with the development of CEN PLUS commands.

'Unit' must be left free for the interface, so the alarm zone is used as
the 'where' and not as 'interface' in UpdateAlarm function.
After this update, the alarm device will have a different ID.
Adding devices to battery status, network status and alarm maintenance.

* OpenWebNet: fix close socket command problem

* OpenWebNet: GEN, AREA and GROUP devices