Update to clean minizip from zlib 1.2.11 and allow dynamic linking
We had a version of contrib/minizip from zlib, that had been partly
C++-ised... and then we weren't even building unzip.cpp because it
had been concatenated to unzip.h and all the functions marked 'inline'.

Kill it with fire, and drop in a clean copy from zlib 1.2.11. Also
make it possible to build with -DUSE_BUILTIN_MINIZIP=NO to use a
system dynamic library instead.